Fix Cyclical Reference By Making Parent Reference Weak
The Dialogue compilation system had a cyclical reference between parent and child that causes the RefCounted to never clean up the resource. This PR decides to make the parent reference weak, which allows RefCounted to be cleaned up by the garbage collector if that's the only remaining reference to a resource.

https://docs.godotengine.org/en/stable/classes/class_refcounted.html

> RefCounted instances caught in a cyclic reference will not be freed automatically. For example, if a node holds a reference to instance A, which directly or indirectly holds a reference back to A, A's reference count will be 2. Destruction of the node will leave A dangling with a reference count of 1, and there will be a memory leak. To prevent this, one of the references in the cycle can be made weak with @GlobalScope.weakref.