Allow disabling automatic update of pot files setting
Previously, translations_pot_files would be updated on save to add new dialogue
files and remove deleted ones.

Some projects may prefer to manage this property manually. In particular,
because the property is serialised by Godot as a single line, parallel branches
which each add dialogue files are guaranteed to cause Git conflicts.

Add a setting to disable this behaviour. Default to the previous behaviour.

https://github.com/nathanhoad/godot_dialogue_manager/issues/887